change to ints

+++ readout/readout.c
@@ -31,7 +31,8 @@
 int sensor_in_ptr = 0;
 int sensor_id[SENSOR_BUFFER_LENGTH];
 uint sensor_cpu_time[SENSOR_BUFFER_LENGTH];
-float sensor_voltage[SENSOR_BUFFER_LENGTH];
+//float sensor_voltage[SENSOR_BUFFER_LENGTH];
+int sensor_voltage[SENSOR_BUFFER_LENGTH];
 
 int zeropoint_in_ptr = 0;
 uint zeropoint_id[ZEROPT_BUFFER_LENGTH];
